Move ClientServiceCompilerEnv to service
