+++ compiler_gym/CMakeLists.txt
@@ -71,7 +71,7 @@ cg_py_library(
   DEPS
     ::validation_error
     ::validation_result
-    compiler_gym::envs::client_service_compiler_env
+    compiler_gym::service::client_service_compiler_env
     compiler_gym::spaces::spaces
     compiler_gym::util::util
   PUBLIC
