+++ compiler_gym/envs/CMakeLists.txt
@@ -11,7 +11,7 @@ cg_py_library(
   SRCS
     "__init__.py"
   DEPS
-    ::client_service_compiler_env
+    ::compiler_env
     compiler_gym::envs::gcc::gcc
     compiler_gym::envs::llvm::llvm
     compiler_gym::envs::loop_tool::loop_tool
@@ -20,30 +20,15 @@ cg_py_library(
 
 cg_py_library(
   NAME
-    client_service_compiler_env
-  SRCS
-    "client_service_compiler_env.py"
-  DEPS
-    ::env_py
-    compiler_gym::compiler_env_state
-    compiler_gym::datasets::datasets
-    compiler_gym::service::service
-    compiler_gym::service::proto::proto
-    compiler_gym::spaces::spaces
-    compiler_gym::util::util
-    compiler_gym::validation_result
-    compiler_gym::views::views
-  PUBLIC
-)
-
-cg_py_library(
-  NAME
-    env_py
+    compiler_env
   SRCS
     "compiler_env.py"
   DEPS
     compiler_gym::spaces::spaces
     compiler_gym::util::util
     compiler_gym::views::views
+    compiler_gym::compiler_env_state
+    compiler_gym::validation_result
+    compiler_gym::datasets::datasets
   PUBLIC
 )
