+++ compiler_gym/envs/__init__.py
@@ -2,7 +2,6 @@
 #
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
-from compiler_gym.envs.client_service_compiler_env import ClientServiceCompilerEnv
 from compiler_gym.envs.compiler_env import CompilerEnv
 from compiler_gym.envs.gcc import GccEnv
 from compiler_gym.envs.llvm.llvm_env import LlvmEnv
@@ -11,7 +10,6 @@ from compiler_gym.util.registration import COMPILER_GYM_ENVS
 
 __all__ = [
     "COMPILER_GYM_ENVS",
-    "ClientServiceCompilerEnv",
     "CompilerEnv",
     "GccEnv",
     "LlvmEnv",
