+++ compiler_gym/envs/gcc/BUILD
@@ -17,9 +17,9 @@ py_library(
     ],
     visibility = ["//visibility:public"],
     deps = [
-        "//compiler_gym/envs:client_service_compiler_env",
         "//compiler_gym/envs/gcc/datasets",
         "//compiler_gym/service",
+        "//compiler_gym/service:client_service_compiler_env",
         "//compiler_gym/service/runtime",  # Implicit dependency of service.
         "//compiler_gym/util",
     ],
