+++ compiler_gym/envs/llvm/BUILD
@@ -48,8 +48,8 @@ py_library(
         ":llvm_benchmark",
         ":llvm_rewards",
         "//compiler_gym/datasets",
-        "//compiler_gym/envs:client_service_compiler_env",
         "//compiler_gym/envs/llvm/datasets",
+        "//compiler_gym/service:client_service_compiler_env",
         "//compiler_gym/spaces",
         "//compiler_gym/third_party/autophase",
         "//compiler_gym/third_party/inst2vec",
