+++ compiler_gym/envs/llvm/llvm_env.py
@@ -11,7 +11,6 @@ from typing import Iterable, List, Optional, Union, cast
 import numpy as np
 
 from compiler_gym.datasets import Benchmark, BenchmarkInitError, Dataset
-from compiler_gym.envs.client_service_compiler_env import ClientServiceCompilerEnv
 from compiler_gym.envs.llvm.datasets import get_llvm_datasets
 from compiler_gym.envs.llvm.llvm_benchmark import ClangInvocation, make_benchmark
 from compiler_gym.envs.llvm.llvm_rewards import (
@@ -19,6 +18,7 @@ from compiler_gym.envs.llvm.llvm_rewards import (
     CostFunctionReward,
     NormalizedReward,
 )
+from compiler_gym.service.client_service_compiler_env import ClientServiceCompilerEnv
 from compiler_gym.spaces import Box, Commandline
 from compiler_gym.spaces import Dict as DictSpace
 from compiler_gym.spaces import Scalar, Sequence
