+++ compiler_gym/envs/loop_tool/CMakeLists.txt
@@ -12,7 +12,7 @@ cg_py_library(
     "loop_tool_env.py"
   DATA compiler_gym::envs::loop_tool::service::service
   DEPS
-    compiler_gym::envs::client_service_compiler_env
+    compiler_gym::service::client_service_compiler_env
     compiler_gym::service::service
     compiler_gym::service::proto::proto
     compiler_gym::service::runtime::runtime
