+++ compiler_gym/envs/loop_tool/loop_tool_env.py
@@ -2,7 +2,7 @@
 #
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
-from compiler_gym.envs.client_service_compiler_env import ClientServiceCompilerEnv
+from compiler_gym.service.client_service_compiler_env import ClientServiceCompilerEnv
 
 
 class LoopToolEnv(ClientServiceCompilerEnv):
