+++ compiler_gym/service/CMakeLists.txt
@@ -13,6 +13,8 @@ cg_py_library(
   DEPS
     ::compilation_session
     ::connection
+    # TODO: add this after circular dependencies are resolved
+    #::client_service_compiler_env
     compiler_gym::service::proto::proto
   PUBLIC
 )
@@ -50,5 +52,23 @@ cg_py_library(
   DEPS
     compiler_gym::service::proto::proto
     compiler_gym::util::util
+  PUBLIC
+)
+
+cg_py_library(
+  NAME
+    client_service_compiler_env
+  SRCS
+    "client_service_compiler_env.py"
+  DEPS
+    compiler_gym::envs::compiler_env
+    compiler_gym::compiler_env_state
+    compiler_gym::datasets::datasets
+    compiler_gym::service::service
+    compiler_gym::service::proto::proto
+    compiler_gym::spaces::spaces
+    compiler_gym::util::util
+    compiler_gym::validation_result
+    compiler_gym::views::views
   PUBLIC
 )
