+++ compiler_gym/spaces/BUILD
@@ -69,7 +69,6 @@ py_library(
     srcs = ["reward.py"],
     deps = [
         ":scalar",
-        "//compiler_gym/service",
         "//compiler_gym/util",
     ],
 )
