+++ examples/example_compiler_gym_service/__init__.py
@@ -83,7 +83,7 @@ class ExampleDataset(Dataset):
 # the example-v0 environment will be available to gym.make(...).
 register(
     id="example-cc-v0",
-    entry_point="compiler_gym.envs:ClientServiceCompilerEnv",
+    entry_point="compiler_gym.service.client_service_compiler_env:ClientServiceCompilerEnv",
     kwargs={
         "service": EXAMPLE_CC_SERVICE_BINARY,
         "rewards": [RuntimeReward()],
@@ -93,7 +93,7 @@ register(
 
 register(
     id="example-py-v0",
-    entry_point="compiler_gym.envs:ClientServiceCompilerEnv",
+    entry_point="compiler_gym.service.client_service_compiler_env:ClientServiceCompilerEnv",
     kwargs={
         "service": EXAMPLE_PY_SERVICE_BINARY,
         "rewards": [RuntimeReward()],
