+++ examples/example_compiler_gym_service/demo_without_bazel.py
@@ -88,7 +88,7 @@ class ExampleDataset(Dataset):
 # Register the environment for use with gym.make(...).
 register(
     id="example-v0",
-    entry_point="compiler_gym.envs:ClientServiceCompilerEnv",
+    entry_point="compiler_gym.service.client_service_compiler_env:ClientServiceCompilerEnv",
     kwargs={
         "service": EXAMPLE_PY_SERVICE_BINARY,
         "rewards": [RuntimeReward()],
