+++ examples/op_benchmarks.py
@@ -28,8 +28,8 @@ import typer
 from tabulate import tabulate
 
 import compiler_gym
+from compiler_gym import CompilerEnv
 from compiler_gym.datasets import BenchmarkInitError
-from compiler_gym.envs.client_service_compiler_env import CompilerEnv
 from compiler_gym.util.executor import Executor
 from compiler_gym.util.logging import init_logging
 from compiler_gym.util.runfiles_path import create_user_logs_dir
